Make DAL function behavior consistent
Each DAL function now returns a value to the caller, representing the
database record (or list of records) on which it operated. The only
exception is the `upsert_resources` function, which is never exposed to
an API endpoint.

If the DAL function fails to identify a record on which to operate,
internal exceptions are suppressed and an empty dict is returned to the
caller. The expectation is that callers will handle the receipt of
empty dictionaries as error scenarios.

Also:
     - Optimize DAL methods to reduce DB query burden
     - Return a 200 status code for successful record deletion
       operations. 204 (NO CONTENT) is unexpected in this scenario.