Improve API Error Handling (#236)
* Make DAL function behavior consistent

Each DAL function now returns a value to the caller, representing the
database record (or list of records) on which it operated. The only
exception is the `upsert_resources` function, which is never exposed to
an API endpoint.

If the DAL function fails to identify a record on which to operate,
internal exceptions are suppressed and an empty dict is returned to the
caller. The expectation is that callers will handle the receipt of
empty dictionaries as error scenarios.

Also:
     - Optimize DAL methods to reduce DB query burden
     - Return a 200 status code for successful record deletion
       operations. 204 (NO CONTENT) is unexpected in this scenario.

* Handle errors raised when operating on DB records that don't exist

The API responds with a status 422 and a useful error message.

* Raise an `HTTPException` for internal server errors

* small fixes for issue 233

* removed a TODO that was fixed by the webserver hitting the database when it starts up

* Respond with a 404 when resources aren't found

* Handle errors when DB queries fail

* Simplify handling when resources aren't found

* remove the sqlalchemy errors and use a 409 error instead

* Cleaner error on uniqueness constraint violations

Better enables programmatic handling of this error type.

* Bail on errors before creating DB sessions

* Return a generic error on failed DB queries

* Move errors to their own file

* Add DatabaseUnavailableError, use it

Co-authored-by: Thomas La Piana <tal103020@icloud.com>