Add MongoDB Support for Array Masking [#129] (#226)
Allow data to be masked within arrays, as well as data within objects within arrays (Nested object erasure support was already introduced here #165). Worth noting, individual elements within arrays are targeted and masked, not the entire array. Support is for MongoDB only, but the most of this work is done in code, and can be used to help support other database implementations in the future.