Independently update `analytics_opt_out` and `analytics_id`
The presence of an `analytics_id` in the config file needs to be
checked explicitly. If a user manually adds `analytics_opt_out =
false`, their `analytics_id` is never added to the config file. This
results in an opted-in user, who continuously pushes events to fideslog
with a newly generated `analytics_id` value.