refactor get_formatted_data_use
This change stemmed from being able to make get_formatted_data_subjects a little more DRY. In the end, the instance types made it somewhat more complicated than a for loop. This should make it easier to add any further attributes in the future, although the special case of legitimate_interest could be further improved in my opinion.