language and guide documentation
This change to the documentation gave me a different viewpoint for the rights available to a data subject. Not only are multiple rights likely to be available for a data subject, but there are some fairly strict categories to be applied that may warrant having a stricter set of values to choose from (for consistency purposes). These changes are meant to reflect the future state to be updated in a follow-on commit.