`fidesctl apply` "upserts" resources (#421)
* `upsert_resources` actually `UPSERT`s resources

And returns the differentiated counts of inserted and updated rows.
These counts are now logged as they once were, long ago.

* Remove redundant logging

* Insert `organization` resources, upsert the rest

If a `default_organization` already exists, it may have been modified
by the user. We don't want to overwrite it with the default taxonomy's
prescribed `default_organization`.

* Reduce `os` import volume

* Log skipped organization inserts

* Add an `/upsert` endpoint for each `resource_type`

* `generate_resource_url` doesn't use the `version`

* `fidesctl apply` uses the `upsert` action

This also removes logging of "skipped" resources, as our implementation
of `upsert` is not smart enough to skip updating resources that are
identical to their already-persisted counterparts.

* Offload unecessary resource parsing to `--dry` operations only

These calculations do not need to be performed for persisting
applications, since the `/upsert` endpoint will return the counts of
inserted vs. updated rows.

This also removes the `unchanged` return value from
`sort_create_update_unchanged`, since it is no longer applicable.

* Fix typos

* Restore `--diff` without `--dry`

* Clarify `apply` log output

It's unclear why a resource that has not change would otherwise output
as `UPDATED`. By removing the distinction between `CREATED` and
`UPDATED` resources during the upsert, we can sidestep this lack of
clarity.

The distinction is still present for `--dry` runs.

* Improve command help text