Model and Export Remaining Privacy Notice Attributes (#413)
* model changes for privacy notice attributes [skip ci]

* validation for legitimate interests

* bump migration, export system, resolve ci checks

* legitimate interests in data map

* export data subject extended attributes

* resequence migrations due to merging

* duplicate dataset.name for privacy attribute

* refactor get_formatted_data_use

This change stemmed from being able to make get_formatted_data_subjects a little more DRY. In the end, the instance types made it somewhat more complicated than a for loop. This should make it easier to add any further attributes in the future, although the special case of legitimate_interest could be further improved in my opinion.

* language and guide documentation

This change to the documentation gave me a different viewpoint for the rights available to a data subject. Not only are multiple rights likely to be available for a data subject, but there are some fairly strict categories to be applied that may warrant having a stricter set of values to choose from (for consistency purposes). These changes are meant to reflect the future state to be updated in a follow-on commit.

* rewrite data_subject rights

The initial issue highlighted a request from a user group around annotating data subject rights. The request centered around templating or optionally only marking the inverse of the data subject rights available. By having the data_subject take both a strategy and optionally a list of rights available as defined by GDPR we can now make this a reality and reduce any potential burden / overhead in assigning rights as part of a data map.

* doc updates for data_subject rights overhaul

* cleanup from self code review

* add new columns to .fides dataset

* remove empty file

* update extending taxonomy docs and references

Co-authored-by: Cole Isaac <cole@ethyca.com>