Add Data Protection Impact Assessment to System Resource Model (#441)
* modeling data protection impact assessments

* export new dpia attributes

This includes default n/a values for the progress and link properties that are not required. Updated testing in this commit as well.

Determined that leaving the progress as a string is likely prefereable to having a strict requirement for the progress status. One user as an example intimiated they would prefer to have that be a link to a the DPIA status. These are likely to evolve over time as users voice opinions on what makes most sense here.

* documentation, tests, ci check fixes

* add missing data_category for new column