data map user guide
This is a first draft of what a user guide might look like. Some further details are still necessary, but I wanted to explore the style of using the demo_resources and extending them appropriately to get to a data map.

Separately, it might make more sense to either tack this on to the flask app walkthrough, create a separate walkthrough, or simplify this by reorganizaing in a meaningful way.