Resolves system and dataset data categories on a data map (#434)
* resolves system and dataset data categories

Out of multiple conversations it was determined that the data categories on a system
should be able to stand alone without a dataset required to create a data map.

Further information highlighted the intended relationship between a system and a
dataset, being the system has the referenced datasets. More clearly, any referenced
datasets will have their data categories applied to the system. This falls in line
with the way evaluations works today and should be mirrored.

With this new information, a deviation from the first planned iteration of just the
data categories on a system seemed prudent. The following change minimizes the
required amount of refactoring by transforming the data frames to support a data map
built of systems with or without a referenced dataset.

In effect, the system is enhanced by the dataset but not required.

* documentation on referenced datasets and data categories

* conclude a comment left full of suspense