review comments
As part of the main commands fix in this PR we needed to ignore some pylint errors for unused imports used strictly for validation. Moving the checks for okta and boto3 into their respective modules was mentioned as being more intutivve and also allowed us to DRY it up with a function for each.