add missing __init__.py (#473)
* add missing __init__.py

This was found shortly after the release of 1.5.0 when trying to run through the fidesdemo. This sort of thing should likely be resolved when we have checks for a pip install of the version as well

* ignore unused import errors for boto3 and okta

* review comments

As part of the main commands fix in this PR we needed to ignore some pylint errors for unused imports used strictly for validation. Moving the checks for okta and boto3 into their respective modules was mentioned as being more intutivve and also allowed us to DRY it up with a function for each.