Orphan Dataset Warning (#543)
* warn of orphan datasets

Due to apply being used as part of an evaluation as well, this seemed like the correct spot to surface the warning most effectively.

The validation compares all referenced datasets from a system against all datasets. If the returned list contains any information a warning is returned. A warning seemed to be a better starting point to alert a user without causing any undue frustration or blockers.

* tests and documentation

* changelog

* typing fix found in review

* Apply suggestions from code review, rename as get_orphan_datasets

Co-authored-by: Thomas <thomas.lapiana+github@pm.me>

Co-authored-by: Thomas <thomas.lapiana+github@pm.me>