audit command and basic check
A new command to audit for compliance seemed to make sense here. Initially I think we may just want to have one or two static checks to help with populating a data map, with some flexibility to add new checks. Down the road, I would like to populate the checks from the server potentially so custom checks can be enabled or disabled as well.