initial commit for a new endpoint
This is still in an in work state, but wanted to push up the changes that I have made so far. Other issues have been opened that will require being merged before this can close. Outside of the initial work for the new endpoint, a new helpers.py has been created to house some reusable items. More to come.