add aws config option, return the systems as json
This is still an iterative state, with the focus being on avoiding any impact to the current cli functionality. One suggestion has been to pass the client around instead of the configuration.

Further work will still be required to separate the cli commands from the core files. This will be done in separate issues.