include default_organization at minimum
This is a near-term workaround until we include all resources on the server as part of the export. This fix will ensure that the default_organization is included at minimum, which is in line with our documentation around being able to utilize default_organization behind the scenes.

This will unblock users from currently hitting an error when trying to utilize an organization not defined in the manifest directory.