Scan/Generate Endpoint - Core Updates (#645)
* add aws config option, return the systems as json

This is still an iterative state, with the focus being on avoiding any impact to the current cli functionality. One suggestion has been to pass the client around instead of the configuration.

Further work will still be required to separate the cli commands from the core files. This will be done in separate issues.

* pulling changes after testing against new endpoint

After testing this PR against the new endpoint, some issues were surfaced with how some core functionality is used. It seemed the best next option was to utilize the specific Core commands required and only update the config option.

Further work in here will look at replacing the config option with generating a session and passing that instead.

* changelog update

* reorganization to embed filter_aws_systems