Export data map from server resources instead of manifest dir (#662)
* All resources for an organization in data map

Previously only fides keys found in a specified manifest directory were included in a data map. This not only led to a bug when an organization is not defined in a manifest directory but also is impractical when working inside a large organization.

This change will have a data map exported for a specific organization, which is more aligned with our documentation around organizations and expactations of users.

* filter resources by organization_fides_key

* change organization key from argument to option

* changelog

* replace manifest_dir arg with option

The positional argument felt a bit out of place. As of today, we can expect most users to have a default_organization and to use the .fides/ directory. These options will allow for deviations from that plan without requiring any specific positional arguments to take place.

* move output-dir option to options.py