Refactor Export to remove CLI specific request methods (#725)
* remove api calls from core functions

This went through a few iterations trying to find the right placement and balance for the formatting of the taxonomy components. In the end moving them as close to their uses seemed to make the most sense. This change should allow for improved reuse of the core functions and helpers that exist as part of exporting resources data from fides.

* changelog

* Apply suggestions from code review

Co-authored-by: Paul Sanders <psanders1@gmail.com>

* black formatting for changes in GitHub

Co-authored-by: Paul Sanders <psanders1@gmail.com>