Handling an invalidly generated fides_key (#761)
* validate the generated fides_key

Due to the myriad of names allowed by different style databases, we need a method that allows some intervention without breaking. This is a start to handle invalid characters when generating a dataset. The function should be allowed to be repurposed when generating other types of resources as well.

A further step would be to also validate if the fides_key already exists due and how we may want to handle that.

* remove manual intervention and cli specific steps

* testing

* rename variable, fix docs from removed manual step