Human-readable column names in datamap response (#779)
* prepend column names to response

This came out of a request to provide human readable names as part of the api response. This also helps resolve a scenario where no data is returned by at minimum providing the column names. This is by no means the end result but a step in the right direction to agreeing a contract between the front end and back end.

* correct return types of datamap route

* duplicate dataset name and sort columns correctly

* remove test logging

* changelog