fidesctl/datamap: Concise heading of Fides Key columns
I noticed that the long description of the fides key column is almost always going
to be truncated in the UI. This is true for other columns, but they all try to be
concise and put most important words first (no "the") which is easier to follow
when hunting for information in a table.

Long-term we'll probably need to introduce a way of embedding help text in this API,
or another API/constants file with the column titles seperate fromt the help text.