cls migration
See [fidesctl-plus#104](https://github.com/ethyca/fidesctl-plus/issues/104) for further rationale, but to summarize the idea of consolidating migrations to a single location was much more popular.

Originally 4 tables were projected, but feedback in the above referenced issue provided insight to the opportunity to consolidate the data from one of the tables. Further changes can be expected to the classification details as more feedback on user experience is understood