cls migration (#1060)
* cls migration

See [fidesctl-plus#104](https://github.com/ethyca/fidesctl-plus/issues/104) for further rationale, but to summarize the idea of consolidating migrations to a single location was much more popular.

Originally 4 tables were projected, but feedback in the above referenced issue provided insight to the opportunity to consolidate the data from one of the tables. Further changes can be expected to the classification details as more feedback on user experience is understood

* replace migration with updates from code review

Some good discussions led to the opportunity to remove the state management as part of the initial deployment as well improved column naming. Additionally, the id being added by default to postgres allowed the remove of some manual column management for our purposes.

Co-authored-by: @PSalant726

* add additional tables to dataset.yml

* updated migration

* reset id columns to defaults

* update instance_id to match id column type

* changelog

* include dataset_key to classify instance table

* include dataset_name as well