ui/system: Improved checking of system create vs editing
DescribeSystemStep was already querying the existing systems. This info can be
used to check if the system that was passed in (if any) already exists (by
fides key). This also disable the fides key field so that the user can't
accidentally switch to _creating_ a new system by modifying that field.