ui/auth: Remove lingering local storage parsing (#1678)
* ui/auth: Remove lingering local storage parsing

I was wondering why we need to manually open up localStorage when we have redux-persist
configured to do that for us now. Removed this and the log in/out still works as
expected. So I think this was just a lingering from changeover.

Similar to this block that was left over from the fidesops merge:
https://github.com/ethyca/fides/pull/1569/files#diff-3071c1012e5bece4441d5b7643c083188c22fbfe96caea78da1f09fdeeeb1599L59

* Update changelog