privacy-center: introduce fides-consent script package
This change introduces a "packages" directory under the Privacy Center, which
will house code that is built to work with the PC's consent options outside
of the PC itself. The first such packages is the "tracking pixel script" which
is a small script exposes the users consent choices through a window global.