incomplete commit to remove load dataset button
Pushing this up and planning to ask for help around how to manage the state so the navigation back works as well. Additionally, the toast message loads twice. This commit attempts to move the `handleLoadDataset` function call to be done on page load if there is an `activeDatasetFidesKey` (essentially what happened on the button click) however I think I am getting lost in some of the state management with dispatch calls from DatasetTable.tsx