Fix docs for `log_pii` config variable
In testing, I found that the docs were inaccurate and this variable is actually read from `FIDES__LOGGING__LOG_PII`, not `FIDES_LOG_PII` (i.e. it follows the same pattern as most other config vars, not the special case "Additional environment variables"). 

I made this change to update the docs to reflect that -- though if it's actually intended that it be a special case env var instead, I can close this PR!