feat: allow icon_path to accept material design and fontawesome icon references
Now, the icon_path fields can accept FontAwesome 5 and Material Design icons instead of just a path to an svg. This maintains compatability with the path-based method by inspecting the value of icon_path. If it starts with md:, it will be a Material Design icon; if it starts with fa:, it will be a FontAwesome icon. See https://react-icons.github.io/react-icons/icons\?name\=md and https://react-icons.github.io/react-icons/icons\?name\=fa for a complete listing of icon names that it accepts.