improve export code to handle privacy declaration custom fields and properly disregard taxonomy custom fields for now
eventually we may want to include taxonomy custom fields in the datamap export, but for now we will disregard them entirely.
previously, they'd been included in the column header row, but their values were not being included, which misaligned the export.
now, they will not be included in the column header row.

also some related cleanup of code to make it a bit easier to follow.