Remove Experience Config Delivery Mechanism [#3322] (#3387)
Removes ExperienceConfig.delivery_mechanism and all associated validation. 

Previously, PrivacyNotices that had opt_in or notice_only consent mechanisms had to be presented in an overlay that was delivered by a banner (field stored on the PrivacyExperienceConfig). There was a lot of behind the scenes activity to ensure that we always kept our PrivacyExperiences in the correct state, when both PrivacyNotices and PrivacyExperienceConfigs were updated. We remove this here and instead surface a property calculated at runtime for the frontend to determine if they need to show a banner.

Similarly makes the related changes to update the ExperienceConfig with the new fields per updated requirements.


Co-authored-by: Allison King <allison@ethyca.com>