[Backend] Further Deduplicate PrivacyExperience and PrivacyExperienceConfig  (#3470)
Deduplicates all the info between PrivacyExperience and PrivacyExperienceConfig tables except for "component" and stop tracking history of privacy experiences now that Privacy Experiences are largely a linking table.

Saving privacy preferences should now pass on the "experience id" instead of the "experience history id" (because it doesn't exist)

Now that we've relaxed a lot of the constraints around privacy experiences we no longer need as many fields on the PrivacyExperience table. There used to be necessary duplication between PrivacyExperience and PrivacyExperienceConfig but now this isn't required and we're just at risk of the tables getting out of sync.