fix: paramaterize exclusion number
Due to differences in the expected strings for an access or an erasure, it was determined that we may need to remove 2 or 3 items from the leading edge of the dataset. To avoid having the same function twice a parameter is now passed based on the string argument passed. The parameter name was also updated to be more descriptive