Refactor how PrivacyPreferenceHistory.determine_relevant_systems works.  This method runs when a privacy preference is saved, and it takes a snapshot of the systems that are relevant for that preference at that point in time.
For TCF systems, a starting point for the system to be relevant is that it has to have a GVL-related data use and a legal basis of processing of consent or legitimate interests.  From there, we further filter those systems that match a specific use, feature, or system/vendor id.