Bugfix: Update the PRIVACY_REQUEST_STATUS_CHANGE_POLL task to check both IN_PROCESSING and APPROVED privacy requests to see if they have failing Request Tasks.
If a Privacy Request needed manual approval failed during processing, it could hang in the "Approved" state, because the polling process wasn't expecting this status.
If a Privacy Request is hanging out in a Pending status it is likely awaiting approval.  Privacy Requests that don't need approval are queued immediately and are transitioned to in-processing.  If a Privacy Request is approved, it is queued and not transitioned to in-processing. Both of these states can indicate a Privacy Request that is in-flight.