Rename Privacy Notice Properties (#5259)
Rename some properties on PrivacyNotice models that have the same name as fields exposed in the API. We want to intentionally choose when these properties are used in the API instead of having them be used automatically. Further, refactor these properties to be slightly more performant when they are called directly.